requirements.txt
BeautifulSoup==3.2.1
Cheetah==2.4.4
Pillow==2.3.0
PyYAML==3.10
XlsxWriter==0.7.7
argparse==1.2.1
backport-collections==0.1
backports.lzma==0.0.3
beautifulsoup4==4.4.1
boto==2.38.0
bottle==0.12.9
bz2file==0.98
cbor==0.1.22
chardet==2.0.1
chromium-compact-language-detector==0.031415
colorama==0.2.5
configobj==4.7.2
coordinate==0.1.5
dblogger==0.4.14
decorator==3.4.0
elasticsearch==1.9.0
enum34==1.1.1
ftfy==4.0.0
futures==3.0.3
gensim==0.12.3
geojson==1.3.1
gevent==1.0
greenlet==0.4.2
happybase==0.9
html5lib==0.999
httpretty==0.8.6
importlib==1.0.3
jellyfish==0.5.1
joblib==0.7.1
jsonpatch==1.3
jsonpointer==1.0
kvlayer==0.5.9
lockfile==0.12.2
logutils==0.3.3
lxml==3.5.0
many-stop-words==0.1.3
matplotlib==1.3.1
names==0.3.0
nilsimsa==0.3.7
nltk==3.0.5
nose==1.3.1
numpy==1.8.2
oauth==1.0.1
pdfminer==20140328
phonenumberslite==7.2.1
prettytable==0.7.2
protobuf==2.6.1
psutil==3.3.0
psycopg2==2.6.1
py==1.4.31
pyaccumulo==1.5.0.8
pyparsing==2.0.1
pyquchk==0.5.3
pyserial==2.6
pytest==2.8.3
pytest-capturelog==0.7
pytest-diffeo==0.1.12
python-daemon==2.1.0
python-dateutil==1.5
python-docx==0.8.5
python-geohash==0.8.5
python-magic==0.4.10
python-snappy==0.5
pytz==2012c
redis==2.10.5
regex==2015.11.22
requests==2.2.1
scikit-learn==0.14.1
scipy==0.13.3
setproctitle==1.1.9
simplejson==3.3.1
six==1.5.2
smart-open==1.3.0
sortedcollection==0.1.11
ssh-import-id==3.21
streamcorpus==0.3.55
streamcorpus-pipeline==0.7.15
tblib==1.1.0
thrift==0.9.3
total-ordering==0.1.0
trollius==2.0
uWSGI==2.0.11.2
urllib3==1.12
urlnorm==1.1.3
virtualenv==1.11.4
wheel==0.24.0
wordsegment==0.6.2
wsgiref==0.1.2
yakonfig==0.7.7
zope.interface==4.0.5
pyopenssl
ndg-httpsclient
pyasn1
certifi
